add io_utils.rs module for saving, loading, and printing
